Some basic cleanups and added new singleton.

+++ ardOinus/src/devel/code/emulate_arduino.cxx
@@ -3,23 +3,24 @@
 
 // Not much of a emulator.  Just allow linking for now.
 
-void pinMode(unsigned pin, std::uint8_t mode) {
+
+void pinMode(int pin, std::uint8_t mode) {
 
 }
 
-void digitalWrite(unsigned pin, std::uint8_t mode) {
+void digitalWrite(int pin, std::uint8_t mode) {
 
 }
 
-int digitalRead(unsigned pin) {
+int digitalRead(int pin) {
   return LOW;
 }
 
-unsigned millis() {
+unsigned long millis() {
   return 1;
 }
 
-unsigned micros() {
+unsigned long micros() {
   return 1000;
 }
